release

+++ deploy/deploy.sh
@@ -1,5 +1,7 @@
 #! /bin/bash
 
+
+# GS_FORCE=1    - Overwrite even if version exists
 VER_RELAY="$(grep AC_INIT ../configure.ac | cut -f3 -d"[" | cut -f1 -d']')"
 VER_GS="$(grep AC_INIT ../../gsocket/configure.ac | cut -f3 -d"[" | cut -f1 -d']')"
 
@@ -19,37 +21,18 @@ deploy_host()
 			(cd gsocket && ./configure --prefix=\$HOME/usr && make all install)"
 	} || exit 255
 
-	ssh -p $port "gsnet@${host}" "[[ \"${VER_RELAY}\" = \$(usr/bin/gsrnd -h 2>&1 | grep ^Vers | cut -f2 -d' ') ]] && exit 0" && { echo "gsocket-relay-${VER_RELAY} already installed. Skipping."; return; }
+	[[ -n $GS_FORCE ]] || { ssh -p $port "gsnet@${host}" "[[ \"${VER_RELAY}\" = \$(usr/bin/gsrnd -h 2>&1 | grep ^Vers | cut -f2 -d' ') ]] && exit 0" && { echo "gsocket-relay-${VER_RELAY} already installed. Skipping."; return; }; }
 	scp -P $port ../gsocket-relay-${VER_RELAY}.tar.gz "gsnet@${host}:"
 	ssh -p $port "gsnet@${host}" "tar xfz gsocket-relay-${VER_RELAY}.tar.gz && \
 		{ [[ -d \$HOME/usr ]] || mkdir \$HOME/usr; } && \
-		{ [[ -f usr/bin/gsrnd ]] && { cp usr/bin/gsrnd usr/bin/gsrnd-\$(usr/bin/gsrnd -h 2>&1 | grep ^Vers | cut -f2 -d' ')_\$(date '+%s'); }; true; } && \
+		{ [[ -f usr/bin/gsrnd ]] && { BDIR=\"backup-\$(usr/bin/gsrnd -h 2>&1 | grep ^Vers | cut -f2 -d' ')_\$(date '+%s')\"; mkdir \"\$BDIR\"; mv usr/bin/gsrnd usr/bin/gsrn_cli \"\$BDIR\"; }; true; } && \
+		{ [[ -h gsocket-relay-${VER_RELAY}/gsocket ]] && rm -f gsocket-relay-${VER_RELAY}/gsocket; true; } && \
 		(cd gsocket-relay-${VER_RELAY} && ln -s ../gsocket gsocket && ./configure --prefix=\$HOME/usr && make all install) && \
 		{ [[ \"${VER_RELAY}\" = \$(usr/bin/gsrnd -h 2>&1 | grep ^Vers | cut -f2 -d' ') ]] || { echo \"${host}: Cant execute gsrnd\"; exit 255; } } && \
 		exit 0
 	"
 }
 
-# deploy_host()
-# {
-# 	local port
-# 	port=$1
-# 	host=$2
-# 	[[ -z $host ]] && { echo >&2 "Bad parameters"; exit 2; } 
-# 	echo "Deploying to ${host}:${port}...."
-# 	ssh -p $port "gsnet@${host}" "[[ \"${VER_RELAY}\" = \$(usr/bin/gsrnd -h 2>&1 | grep ^Vers | cut -f2 -d' ') ]] && exit 0" && { echo "gsocket-relay-${VER_RELAY} already installed. Skipping."; return; }
-# 	scp -P $port ../../gsocket/gsocket-${VER_GS}.tar.gz ../gsocket-relay-${VER_RELAY}.tar.gz "gsnet@${host}:"
-# 	ssh -p $port "gsnet@${host}" "tar xfz gsocket-${VER_GS}.tar.gz && \
-# 	tar xfz gsocket-relay-${VER_RELAY}.tar.gz && \
-# 	ln -sf gsocket-${VER_GS} gsocket && \
-# 	{ [[ -d \$HOME/usr ]] || mkdir \$HOME/usr; } && \
-# 	(cd gsocket && ./configure --prefix=\$HOME/usr && make all install) && \
-# 	[[ -f usr/bin/gsrnd ]] && { cp usr/bin/gsrnd usr/bin/gsrnd-\$(usr/bin/gsrnd -h 2>&1 | grep ^Vers | cut -f2 -d' ')_\$(date '+%s'); } && \
-# 	(cd gsocket-relay-${VER_RELAY} && ./configure --prefix=\$HOME/usr && make all install) && \
-# 	{ [[ \"${VER_RELAY}\" = \$(usr/bin/gsrnd -h 2>&1 | grep ^Vers | cut -f2 -d' ') ]] || { echo \"${host}: Cant execute gsrnd\"; exit 255; } } && \
-# 	exit 0
-# 	"
-# }
 
 restart_host()
 {
